pyproject.toml
[tool.poetry]
name = "osiq"
version = "0.1.0"
description = ""
authors = ["Adnantium <adnantium@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.23.0"
pylint = "^3.1.0"
pylint-django = "^2.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
